Add files via upload
Download this compressed ML-DATA-PIPELINE-MAIN.zip file and extract in you local system to run the pipeline.  Because in original uploaded file, may be some documents will not get uploaded. For that, download this alternate same ML-DATA-PIPELINE-MAIN.zip  file and extract in your system and follow the commands one by one that are mentioned in README.md file(You can see in original uploaded file) to execute the pipeline.